+++ changes.md
@@ -7,3 +7,11 @@
 - :warning: changed property `column` of type `GraphQLErrorLocation` from type `integer` to `number`
 </details>
 
+
+<details>
+<summary>Added QueryParameter(s)</summary>
+
+- added query parameter `where` to method `get /{projectKey}/product-selections/key={key}/products`
+- added query parameter `where` to method `get /{projectKey}/product-selections/{ID}/products`
+</details>
+